# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example for response examples value",
        "version": "1.0"
    },
    "paths": {
...
        "/api/prices": {
            "post": {
                "tags": [
                    "Prices"
                ],
                "summary": "Create new",
                "operationId": "dde05fafa147fe190eda2eafb1d41891",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "configuration_id": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "integer"
                                    },
                                    "start_date": {
                                        "type": "string"
                                    },
                                    "end_date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "configuration_id": "10",
                                    "price": 10000,
                                    "start_date": "2020-09-11 12:12:12",
                                    "end_date": "2020-09-12 12:12:12"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Price"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/prices/{priceId}": {
            "get": {
                "tags": [
                    "Prices"
                ],
                "summary": "Get by ID",
                "operationId": "f16638207f70090890d0ba980cd519da",
                "parameters": [
                    {
                        "name": "priceId",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Price"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Prices"
                ],
                "summary": "Delete by ID",
                "operationId": "7b882d2f28355aae3dc1fe7c16db1ae5",
                "parameters": [
                    {
                        "name": "priceId",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully"
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Prices"
                ],
                "summary": "Update by ID",
                "operationId": "554d11c6cf68742d64ddfafbe7405454",
                "parameters": [
                    {
                        "name": "priceId",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "configuration_id",
                                    "price",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "configuration_id": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "1000"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "example": "2020-09-11 12:12:12"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "example": "2020-09-12 12:12:12"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "required": [
                    "name"
                ],
                "properties": {
                    "code": {
                        "description": "Code",
                        "type": "integer",
                        "example": 404
                    },
                    "message": {
                        "description": "Message",
                        "type": "string",
                        "example": "Not found"
                    }
                },
                "type": "object"
            },
...
            "Price": {
                "required": [
                    "configuration_id",
                    "price",
                    "start_date",
                    "end_date"
                ],
                "properties": {
                    "configuration_id": {
                        "description": "Configuration ID",
                        "type": "integer",
                        "example": 10
                    },
                    "price": {
                        "description": "Price",
                        "type": "integer",
                        "example": 1000
                    },
                    "start_date": {
                        "description": "Start date",
                        "type": "string",
                        "example": "2020-09-11 12:12:12"
                    },
                    "end_date": {
                        "description": "End date",
                        "type": "string",
                        "example": "2020-09-12 12:12:12"
                    }
                },
                "type": "object"
            }
        }
    },
...
}